code/pg_io_predictor/requirements.txt
pyyaml>=6.0
psycopg2-binary>=2.9
sqlparse>=0.4
pytest>=7.0
tabulate>=0.9
